% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortPercent.R
\name{CohortPercent}
\alias{CohortPercent}
\title{Functions to help reporting churn}
\usage{
CohortPercent(df, c, var)
}
\arguments{
\item{c}{}
}
\value{
A bar graph describing churn patterns, discriminated by c.
}
\description{
function calculates the churn cohortwise
}
\examples{

}
