% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeFun.R
\name{completeFun}
\alias{completeFun}
\title{Functions to help reporting churn}
\usage{
completeFun(data, desiredCols)
}
\arguments{
\item{df, }{c, var.}
}
\value{
Filtered data, without na on certain columns
}
\description{
Enable na deletion based in a given set of columns
}
\examples{

}
