% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempoVida_Hist.R
\name{TempoVida_Hist}
\alias{TempoVida_Hist}
\title{Functions to help reporting churn}
\usage{
TempoVida_Hist(df, c, var)
}
\arguments{
\item{df, }{c, var.}
}
\value{
An histogram, discriminated by c.
}
\description{
They let knitr script looks cleaner
}
\examples{

}
